Personal AI Agent/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.11.5
langchain==0.3.25
faiss-cpu==1.11.0
sentence-transformers==4.1.0
llama-cpp-python==0.3.9
python-multipart==0.0.20
aiofiles==24.1.0
tqdm==4.67.1
requests==2.32.4
httpx==0.28.1
python-dotenv==1.1.0
openpyxl==3.1.5
pypdf==5.6.1
python-magic==0.4.27